# backend/src/main/java/com/contacthub/api/CategoryController.java
package com.contacthub.api;

import com.contacthub.domain.CategoryEntity;
import com.contacthub.service.CategoryService;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.UUID;

@RestController
@RequestMapping("/api/categories")
public class CategoryController {
    
    private final CategoryService categoryService;
    
    public CategoryController(CategoryService categoryService) {
        this.categoryService = categoryService;
    }
    
    @GetMapping
    public ResponseEntity<List<CategoryEntity>> getAllCategories() {
        return ResponseEntity.ok(categoryService.getAllCategories());
    }
    
    @GetMapping("/{id}")
    public ResponseEntity<CategoryEntity> getCategoryById(@PathVariable UUID id) {
        return categoryService.getCategoryById(id)
            .map(ResponseEntity::ok)
            .orElse(ResponseEntity.notFound().build());
    }
    
    @PostMapping
    public ResponseEntity<CategoryEntity> createCategory(@RequestBody CategoryRequest request) {
        CategoryEntity category = categoryService.createCategory(request.name(), request.color());
        return ResponseEntity.status(HttpStatus.CREATED).body(category);
    }
    
    @PutMapping("/{id}")
    public ResponseEntity<CategoryEntity> updateCategory(
            @PathVariable UUID id,
            @RequestBody CategoryRequest request) {
        return categoryService.updateCategory(id, request.name(), request.color())
            .map(ResponseEntity::ok)
            .orElse(ResponseEntity.notFound().build());
    }
    
    @DeleteMapping("/{id}")
    public ResponseEntity<?> deleteCategory(@PathVariable UUID id) {
        boolean deleted = categoryService.deleteCategory(id);
        return deleted 
            ? ResponseEntity.noContent().<ResponseEntity<?>>build() 
            : ResponseEntity.notFound().<ResponseEntity<?>>build();
    }
    
    public record CategoryRequest(String name, String color) {}
}