# specs/001-contacthub/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: ContactHub API
  description: Contact Management REST API
  version: 1.0.0

servers:
  - url: http://localhost:8080/api
    description: Development server

paths:
  /contacts:
    get:
      summary: List all contacts
      description: Returns a paginated list of contacts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: tagId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedContacts'
    post:
      summary: Create a new contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContactRequest'
      responses:
        '201':
          description: Contact created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'

  /contacts/{id}:
    get:
      summary: Get contact by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Contact found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '404':
          description: Contact not found
    put:
      summary: Update contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContactRequest'
      responses:
        '200':
          description: Contact updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '404':
          description: Contact not found
    delete:
      summary: Delete contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Contact deleted
        '404':
          description: Contact not found

  /contacts/search:
    get:
      summary: Search contacts
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contact'

  /categories:
    get:
      summary: List all categories
      responses:
        '200':
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
    post:
      summary: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'

  /categories/{id}:
    get:
      summary: Get category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Category found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          description: Category not found
    put:
      summary: Update category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryRequest'
      responses:
        '200':
          description: Category updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          description: Category not found
    delete:
      summary: Delete category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Category deleted
        '404':
          description: Category not found

  /tags:
    get:
      summary: List all tags
      responses:
        '200':
          description: Tags list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
    post:
      summary: Create a new tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTagRequest'
      responses:
        '201':
          description: Tag created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'

  /tags/{id}:
    get:
      summary: Get tag by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tag found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '404':
          description: Tag not found
    put:
      summary: Update tag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTagRequest'
      responses:
        '200':
          description: Tag updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '404':
          description: Tag not found
    delete:
      summary: Delete tag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Tag deleted
        '404':
          description: Tag not found

components:
  schemas:
    Contact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        notes:
          type: string
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateContactRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 255
        phone:
          type: string
          maxLength: 50
        email:
          type: string
          maxLength: 255
        notes:
          type: string
        categoryIds:
          type: array
          items:
            type: string
            format: uuid
        tagIds:
          type: array
          items:
            type: string
            format: uuid

    UpdateContactRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        phone:
          type: string
          maxLength: 50
        email:
          type: string
          maxLength: 255
        notes:
          type: string
        categoryIds:
          type: array
          items:
            type: string
            format: uuid
        tagIds:
          type: array
          items:
            type: string
            format: uuid

    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        color:
          type: string
        createdAt:
          type: string
          format: date-time

    CreateCategoryRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 100
        color:
          type: string
          maxLength: 20

    UpdateCategoryRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        color:
          type: string
          maxLength: 20

    Tag:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time

    CreateTagRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 50

    UpdateTagRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 50

    PagedContacts:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Contact'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer